Fix controller paths after crate split
