--- controller/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "controller"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-model = { workspace = true }
-view = { workspace = true }
-actix-web = { workspace = true }
-serde = { workspace = true }
-sqlx = { workspace = true }
