--- controller/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod order;
-pub mod root;
-pub mod todo;
