--- controller/src/root.rs
@@ -1,9 +1,0 @@
-use view::html;
-use actix_web::{get, HttpResponse, Responder};
-
-#[get("/")]
-async fn index() -> impl Responder {
-    HttpResponse::Ok()
-        .content_type("text/html")
-        .body(html::root::render_index_page())
-}
