--- model/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod order;
-pub mod todo;
