+++ order/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod model;
+pub mod repository;
+pub mod controller;
+pub mod view;
