--- repository/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod postgres_todo_repository;
-pub mod postgres_order_repository;
