+++ todo/Cargo.toml
@@ -1,13 +1,15 @@
 [package]
-name = "model"
+name = "todo"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-sqlx = { workspace = true }
-serde = { workspace = true }
 chrono = { workspace = true }
+serde = { workspace = true }
+sqlx = { workspace = true }
 async-trait = { workspace = true }
+actix-web = { workspace = true }
+tera = "1"
 
 [dev-dependencies]
 tokio = { workspace = true }
