--- view/templates/index.html
@@ -1,8 +1,0 @@
-{% extends 'layout.html' %}
-{% block content %}
-<section class="hero">
-  <div class="hero-body">
-    <p class="title">Welcome to my app!</p>
-  </div>
-</section>
-{% endblock %}
